# src/main/resources/api/order_openapi.yml
openapi: 3.0.3
info:
  title: Order Service API
  description: API for managing orders
  version: 1.0.0
servers:
  - url: http://localhost:8080/ms-order
paths:
...
  /carts/by-user/{user_id}:
    get:
      summary: Get the cart for a specific user
      description: Retrieve the cart associated with the specified user ID. If no cart exists, returns a `204 No Content` status.
      tags:
        - Cart
      operationId: getCartsByUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The cart associated with the user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '204':
          description: No cart found for the specified user.
  /carts/add-product/{user_id}:
    post:
      summary: Add product to cart
      tags:
        - Cart
      operationId: addProductToCart
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose cart the product will be added to.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product added to the cart successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: User ID and Product must not be null.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User ID and Product must not be null.
                  timestamp:
                    type: string
                    format: date-time
                    example: "YYYY-MM-DDThh:mm:ss.SSS"
        '404':
          description: Product with ID {id} does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Product with ID {id} does not exist.
                  timestamp:
                    type: string
                    format: date-time
                    example: "YYYY-MM-DDThh:mm:ss.SSS"
        '500':
          description: Failed to save the cart.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to save the cart.
                  timestamp:
                    type: string
                    format: date-time
                    example: "YYYY-MM-DDThh:mm:ss.SSS"
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
    Cart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: string
        createdDate:
          type: string
          format: date-time
        cartProducts:
          type: array
          items:
            $ref: '#/components/schemas/CartProduct'
    CartProduct:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          $ref: '#/components/schemas/Product'
        productQuantity:
          type: integer
          format: int64
...